experimented with bias variance, dimensinaity  reduction (PCA) with multiple models:
1. Linear Models (Ridge, Lasso, ElasticNet)
2. Tree-Based Models (Decision Tree & Random Forest)
3. Neural Network (MLP)

Added a step by step guide to the notebook in the shape of comments and markdown explaining what im doing and why. So both the reader and myself have a good understand of each step.

TO:DO write a README for this experiment and dive deeper into this dataset and make a notebook with the best models/ settings to predict the test data.